Module actix_web::actix::actix

Modules

dev
The actix prelude for library developers

io

Structs

Addr
Address of the actor

Condition

Context
Actor execution context

MessageResult
Helper type that implements MessageResponse trait

Supervisor
Actor supervisor

SyncArbiter
Sync arbiter

Enums

ActorState
Actor execution state

MailboxError
Set of error that can occurred during message delivery process

Running

SendError

Traits

Actor
Actors are objects which encapsulate state and behavior.

ContextFutureSpawner
Helper trait which can spawn future into actor's context

Supervised
Actors with ability to restart after failure

SystemService
Trait defines system's service.

WrapFuture
Helper trait that allows conversion of normal future into ActorFuture

Type Definitions

ResponseActFuture
A specialized actor future for async message handler