Struct actix_web::actix::actix::dev::Registry

pub struct Registry { /* fields omitted */ }

Actors registry

Actor can register itself as a service. Service can be defined as ArbiterService which is unique per arbiter or SystemService which is unique per system.

If arbiter service is used outside of runnig arbiter, it panics.

Example

use actix::prelude::*;

#[derive(Message)]
struct Ping;

#[derive(Default)]
struct MyActor1;

impl Actor for MyActor1 {
    type Context = Context<Self>;
}
impl actix::Supervised for MyActor1 {}

impl ArbiterService for MyActor1 {
   fn service_started(&mut self, ctx: &mut Context<Self>) {
      println!("Service started");
   }
}

impl Handler<Ping> for MyActor1 {
   type Result = ();

   fn handle(&mut self, _: Ping, ctx: &mut Context<Self>) {
      println!("ping");
   }
}

struct MyActor2;

impl Actor for MyActor2 {
   type Context = Context<Self>;

   fn started(&mut self, _: &mut Context<Self>) {
      // get MyActor1 addres from the registry
      let act = Arbiter::registry().get::<MyActor1>();
      act.do_send(Ping);
   }
}

fn main() {
    // initialize system
    let code = System::run(|| {

        // Start MyActor1 in new Arbiter
        Arbiter::start(|_| {
            MyActor2
        });
    });
}

Methods

impl Registry

pub fn get<A>(&self) -> Addr<A> where
A: Actor<Context = Context<A>> + ArbiterService,

Query registry for specific actor. Returns address of the actor. If actor is not registered, starts new actor and return address of newly created actor.

pub fn set<A>(&self, addr: Addr<A>) where
A: Actor<Context = Context<A>> + ArbiterService,

Add new actor to the registry by address, panic if actor is already running