Trait actix_web::actix::fut::WrapFuture

pub trait WrapFuture<A> where
    A: Actor,
    <Self::Future as ActorFuture>::Item == Self::Item,
    <Self::Future as ActorFuture>::Error == Self::Error,
    <Self::Future as ActorFuture>::Actor == A, {
    type Future: ActorFuture;
    type Item;
    type Error;
    fn into_actor(self, a: &A) -> Self::Future;
}

Helper trait that allows conversion of normal future into ActorFuture

Associated Types

type Future: ActorFuture

The future that this type can be converted into.

type Item

The item that the future may resolve with.

type Error

The error that the future may resolve with.

Required Methods

fn into_actor(self, a: &A) -> Self::Future

Convert normal future to a ActorFuture