Struct actix_web::dev::ResourceInfo

pub struct ResourceInfo { /* fields omitted */ }

Information about current resource

Methods

impl ResourceInfo

pub fn name(&self) -> &str

Name os the resource

pub fn rdef(&self) -> Option<&ResourceDef>

This method returns reference to matched ResourceDef object.

pub fn match_info(&self) -> &Params

Get a reference to the Params object.

Params is a container for url parameters. A variable segment is specified in the form {identifier}, where the identifier can be used later in a request handler to access the matched value for that segment.

pub fn url_for<U, I>(
&self,
req: &Request,
name: &str,
elements: U
) -> Result<Url, UrlGenerationError> where
U: IntoIterator<Item = I>,
I: AsRef<str>,

Generate url for named resource

Check [HttpRequest::url_for()](../struct.HttpRequest.html#method. url_for) for detailed information.

pub fn has_resource(&self, path: &str) -> bool

Check if application contains matching resource.

This method does not take prefix into account. For example if prefix is /test and router contains route /name, following path would be recognizable /test/name but has_resource() call would return false.

pub fn has_prefixed_resource(&self, path: &str) -> bool

Check if application contains matching resource.

This method does take prefix into account but behaves like has_route in case prefix is not set in the router.

For example if prefix is /test and router contains route /name, the following path would be recognizable /test/name and has_prefixed_route() call would return true. It will not match against prefix in case it's not given. For example for /name with a /test prefix would return false