Enum actix_web::http::header::Entry

pub enum Entry<'a, T> where
    T: 'a,  {
    Occupied(OccupiedEntry<'a, T>),
    Vacant(VacantEntry<'a, T>),
}

A view into a single location in a HeaderMap, which may be vacant or occupied.

Variants
Occupied(OccupiedEntry<'a, T>)

An occupied entry

Vacant(VacantEntry<'a, T>)

A vacant entry

Methods

impl<'a, T> Entry<'a, T>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn or_insert(self, default: T) -> &'a mut T

Ensures a value is in the entry by inserting the default if empty.

Returns a mutable reference to the first value in the entry.

Examples

let mut map: HeaderMap<u32> = HeaderMap::default();

let headers = &[
    "content-length",
    "x-hello",
    "Content-Length",
    "x-world",
];

for &header in headers {
    let counter = map.entry(header)
        .expect("valid header names")
        .or_insert(0);
    *counter += 1;
}

assert_eq!(map["content-length"], 2);
assert_eq!(map["x-hello"], 1);

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn or_insert_with<F>(self, default: F) -> &'a mut T where
F: FnOnce() -> T,

Ensures a value is in the entry by inserting the result of the default function if empty.

The default function is not called if the entry exists in the map. Returns a mutable reference to the first value in the entry.

Examples

Basic usage.

let mut map = HeaderMap::new();

let res = map.entry("x-hello").unwrap()
    .or_insert_with(|| "world".parse().unwrap());

assert_eq!(res, "world");

The default function is not called if the entry exists in the map.

let mut map = HeaderMap::new();
map.insert(HOST, "world".parse().unwrap());

let res = map.entry("host")
    .expect("host is a valid string")
    .or_insert_with(|| unreachable!());

assert_eq!(res, "world");

pub fn key(&self) -> &HeaderName

Returns a reference to the entry's key

Examples

let mut map = HeaderMap::new();

assert_eq!(map.entry("x-hello").unwrap().key(), "x-hello");