Struct actix_web::http::header::VacantEntry

pub struct VacantEntry<'a, T> where
    T: 'a,  { /* fields omitted */ }

A view into a single empty location in a HeaderMap.

This struct is returned as part of the Entry enum.

Methods

impl<'a, T> VacantEntry<'a, T>

pub fn key(&self) -> &HeaderName

Returns a reference to the entry's key

Examples

let mut map = HeaderMap::new();

assert_eq!(map.entry("x-hello").unwrap().key().as_str(), "x-hello");

pub fn into_key(self) -> HeaderName

Take ownership of the key

Examples

let mut map = HeaderMap::new();

if let Entry::Vacant(v) = map.entry("x-hello").unwrap() {
    assert_eq!(v.into_key().as_str(), "x-hello");
}

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn insert(self, value: T) -> &'a mut T

Insert the value into the entry.

The value will be associated with this entry's key. A mutable reference to the inserted value will be returned.

Examples

let mut map = HeaderMap::new();

if let Entry::Vacant(v) = map.entry("x-hello").unwrap() {
    v.insert("world".parse().unwrap());
}

assert_eq!(map["x-hello"], "world");

pub fn insert_entry(self, value: T) -> OccupiedEntry<'a, T>

Insert the value into the entry.

The value will be associated with this entry's key. The new OccupiedEntry is returned, allowing for further manipulation.

Examples

let mut map = HeaderMap::new();

if let Entry::Vacant(v) = map.entry("x-hello").unwrap() {
    let mut e = v.insert_entry("world".parse().unwrap());
    e.insert("world2".parse().unwrap());
}

assert_eq!(map["x-hello"], "world2");