Struct actix_web::middleware::identity::CookieIdentityPolicy

pub struct CookieIdentityPolicy(_);

Use cookies for request identity storage.

The constructors take a key as an argument. This is the private key for cookie - when this value is changed, all identities are lost. The constructors will panic if the key is less than 32 bytes in length.

Example

use actix_web::middleware::identity::{CookieIdentityPolicy, IdentityService};
use actix_web::App;

fn main() {
    let app = App::new().middleware(IdentityService::new(
        // <- create identity middleware
        CookieIdentityPolicy::new(&[0; 32])  // <- construct cookie policy
               .domain("www.rust-lang.org")
               .name("actix_auth")
               .path("/")
               .secure(true),
    ));
}

Methods

impl CookieIdentityPolicy

pub fn new(key: &[u8]) -> CookieIdentityPolicy

Construct new CookieIdentityPolicy instance.

Panics if key length is less than 32 bytes.

pub fn path<S: Into<String>>(self, value: S) -> CookieIdentityPolicy

Sets the path field in the session cookie being built.

pub fn name<S: Into<String>>(self, value: S) -> CookieIdentityPolicy

Sets the name field in the session cookie being built.

pub fn domain<S: Into<String>>(self, value: S) -> CookieIdentityPolicy

Sets the domain field in the session cookie being built.

pub fn secure(self, value: bool) -> CookieIdentityPolicy

Sets the secure field in the session cookie being built.

If the secure field is set, a cookie will only be transmitted when the connection is secure - i.e. https

pub fn max_age(self, value: Duration) -> CookieIdentityPolicy

Sets the max-age field in the session cookie being built.